+++ tests/test_types.py
@@ -3,7 +3,7 @@ import unittest
 class TypesTest(unittest.TestCase):
     def test_import(self):
         try:
-            from coretypes import Frame, FrameType, SecurityType, MarketType, bars_with_limit_dtype, bars_dtype
+            from coretypes import Frame, FrameType, SecurityType, MarketType, bars_dtype, bars_dtype_with_code, bars_cols, bars_with_limit_dtype, bars_with_limit_cols, BarsArray, BarsWithLimitArray, BarsPanel, security_db_dtype, security_info_dtype,xrxd_info_dtype
 
             from coretypes import QuotesFetcher
         except Exception: