+++ HISTORY.md
@@ -1,5 +1,10 @@
 # History
 
+## 0.5.0 （2022-08-06）
+
+* 增加了BarsArray, BarsPanel和BarsWithLimitArray数据类型
+* security_db_dtype, security_info_dtype, xrxd_info_dtype
+
 ## 0.1.0 (2022-01-12)
 
 * First release on PyPI.